move public spl transfer type to its own file
